refactor(runtime): one gated call for the acceleration-block notices, one shared log-capture test helper
Folds `warn_about_discarded_acceleration_settings` and the new
`warn_about_deprecated_ready_state` into `warn_about_acceleration_block`, so
each load-path call site gates on `log_errors` once and the next such notice
does not add a third copy at both sites. Behaviour is unchanged.

The regression test's tracing-capture writer was the crate's fourth private
copy of the same twenty lines; `tracing_util::warn_lines_emitted_by` is the
shared `#[cfg(test)]` one. Existing copies are left in place.
