+++ crates/runtime/src/component/mod.rs
@@ -85,9 +85,8 @@ impl AcceleratedComponent {
 }
 
 /// The warning that a component set `acceleration.ready_state`, which is honoured
-/// but deprecated in favour of the component's own top-level `ready_state`. The
-/// load path emits it, once per component
-/// (`init::dataset::warn_about_deprecated_ready_state`).
+/// but deprecated in favour of the component's own top-level `ready_state`. Emitted
+/// by the load path (`init::dataset::warn_about_acceleration_block`).
 ///
 /// One function for both components because it is one deprecation of one key: a
 /// dataset and a view that write the same thing should be told the same thing,
